Move ErrorResponseDecorator
