+++ joylive-plugin/joylive-router/joylive-router-springgateway4/src/main/java/com/jd/live/agent/plugin/router/springgateway/v4/cluster/GatewayCluster.java
@@ -27,31 +27,16 @@ import com.jd.live.agent.plugin.router.springcloud.v4.cluster.AbstractCloudClust
 import com.jd.live.agent.plugin.router.springgateway.v4.cluster.context.GatewayClusterContext;
 import com.jd.live.agent.plugin.router.springgateway.v4.filter.LiveGatewayFilterChain;
 import com.jd.live.agent.plugin.router.springgateway.v4.request.GatewayCloudClusterRequest;
+import com.jd.live.agent.plugin.router.springgateway.v4.response.ErrorResponseDecorator;
 import com.jd.live.agent.plugin.router.springgateway.v4.response.GatewayClusterResponse;
 import lombok.Getter;
-import org.reactivestreams.Publisher;
 import org.springframework.cloud.client.ServiceInstance;
 import org.springframework.cloud.client.loadbalancer.reactive.ReactiveLoadBalancer;
 import org.springframework.cloud.gateway.filter.GatewayFilterChain;
-import org.springframework.core.io.buffer.DataBuffer;
-import org.springframework.core.io.buffer.DataBufferFactory;
-import org.springframework.core.io.buffer.DataBufferUtils;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
-import org.springframework.http.server.reactive.ServerHttpResponse;
-import org.springframework.http.server.reactive.ServerHttpResponseDecorator;
-import org.springframework.lang.NonNull;
 import org.springframework.web.server.ServerWebExchange;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
 
-import java.nio.charset.StandardCharsets;
-import java.util.List;
 import java.util.Set;
 import java.util.concurrent.CompletionStage;
-
-import static org.springframework.cloud.gateway.support.ServerWebExchangeUtils.ORIGINAL_RESPONSE_CONTENT_TYPE_ATTR;
 
 @Getter
 public class GatewayCluster extends AbstractCloudCluster<
@@ -68,7 +53,7 @@ public class GatewayCluster extends AbstractCloudCluster<
         try {
             Set<ErrorPolicy> policies = request.removeErrorPolicies();
             // decorate response to remove exception header and get body
-            BodyResponseDecorator decorator = new BodyResponseDecorator(request.getExchange(), policies);
+            ErrorResponseDecorator decorator = new ErrorResponseDecorator(request.getExchange(), policies);
             ServerWebExchange exchange = request.getExchange().mutate().response(decorator).build();
             GatewayClusterResponse response = new GatewayClusterResponse(exchange.getResponse(),
                     () -> (ServiceError) exchange.getAttributes().remove(Request.KEY_SERVER_ERROR),
@@ -97,71 +82,6 @@ public class GatewayCluster extends AbstractCloudCluster<
     @Override
     protected GatewayClusterResponse createResponse(ServiceError error, ErrorPredicate predicate) {
         return new GatewayClusterResponse(error, predicate);
-    }
-
-    /**
-     * A decorator for {@link ServerHttpResponse} that modifies the response body according to the specified code policies.
-     */
-    private static class BodyResponseDecorator extends ServerHttpResponseDecorator {
-
-        private final ServerWebExchange exchange;
-
-        private final Set<ErrorPolicy> policies;
-
-        BodyResponseDecorator(ServerWebExchange exchange, Set<ErrorPolicy> policies) {
-            super(exchange.getResponse());
-            this.exchange = exchange;
-            this.policies = policies;
-        }
-
-        @NonNull
-        @Override
-        public Mono<Void> writeWith(@NonNull Publisher<? extends DataBuffer> body) {
-            final HttpHeaders headers = exchange.getResponse().getHeaders();
-            ServiceError error = ServiceError.build(key -> {
-                List<String> values = headers.remove(key);
-                return values == null || values.isEmpty() ? null : values.get(0);
-            });
-            if (error != null) {
-                exchange.getAttributes().put(Request.KEY_SERVER_ERROR, error);
-            }
-            if (policies != null && !policies.isEmpty()) {
-                String contentType = exchange.getAttribute(ORIGINAL_RESPONSE_CONTENT_TYPE_ATTR);
-                if (body instanceof Flux && policyMatch(contentType)) {
-                    Flux<? extends DataBuffer> fluxBody = Flux.from(body);
-                    return super.writeWith(fluxBody.buffer().map(dataBuffers -> {
-                        DataBufferFactory bufferFactory = bufferFactory();
-                        DataBuffer join = bufferFactory.join(dataBuffers);
-                        byte[] content = new byte[join.readableByteCount()];
-                        join.read(content);
-                        // must release the buffer
-                        DataBufferUtils.release(join);
-                        exchange.getAttributes().put(Request.KEY_RESPONSE_BODY, new String(content, StandardCharsets.UTF_8));
-                        return bufferFactory.wrap(content);
-                    }));
-                }
-            }
-            return super.writeWith(body);
-        }
-
-        /**
-         * Checks if any of the code policies match the given content type.
-         *
-         * @param contentType The content type to check.
-         * @return true if any of the code policies match the content type, false otherwise.
-         */
-        private boolean policyMatch(String contentType) {
-            contentType = contentType == null ? null : contentType.toLowerCase();
-            HttpStatusCode statusCode = getStatusCode();
-            Integer status = statusCode == null ? null : statusCode.value();
-            int ok = HttpStatus.OK.value();
-            for (ErrorPolicy policy : policies) {
-                if (policy.match(status, contentType, ok)) {
-                    return true;
-                }
-            }
-            return false;
-        }
     }
 
 }
